+++ CHANGELOG.md
@@ -11,6 +11,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/), and this
 ### Added
 
 - `is:command` filter to search for snippets with a command name set
+- Archive/unarchive your own snippets
+- `is:archived` filter to search for archived snippets
+    - By default, archived snippets do not show up in search results
 
 
 ## [1.0.0] - 2025-02-05